+++ src/tmtccmd/com_if/dummy_com_if.py
@@ -2,7 +2,8 @@
 """
 from tmtccmd.com_if.com_interface_base import CommunicationInterface
 from spacepackets.ecss.tc import PusTelecommand
-from tmtccmd.tm import TelemetryListT, Service1TM
+from tmtccmd.tm import TelemetryListT
+from tmtccmd.pus.service_1_verification import Service1TMExtended
 from tmtccmd.pus.service_17_test import Srv17Subservices, Service17TMExtended
 from tmtccmd.utility.logger import get_console_logger
 from tmtccmd.utility.tmtc_printer import TmTcPrinter
@@ -70,7 +71,7 @@ class DummyHandler:
         """
         if self.last_service == 17:
             if self.last_subservice == 1:
-                tm_packer = Service1TM(
+                tm_packer = Service1TMExtended(
                     subservice=1, ssc=self.current_ssc, tc_packet_id=self.last_tc_packet_id,
                     tc_ssc=self.last_tc_ssc
                 )
@@ -78,7 +79,7 @@ class DummyHandler:
                 self.current_ssc += 1
                 tm_packet_raw = tm_packer.pack()
                 self.next_telemetry_package.append(tm_packet_raw)
-                tm_packer = Service1TM(
+                tm_packer = Service1TMExtended(
                     subservice=7, ssc=self.current_ssc, tc_packet_id=self.last_tc_packet_id,
                     tc_ssc=self.last_tc_ssc
                 )
