+++ src/tmtccmd/tm/__init__.py
@@ -1,8 +1,5 @@
-from tmtccmd.tm.definitions import TelemetryListT, TelemetryQueueT
-
-from tmtccmd.tm.service_1_verification import Service1TM
-from tmtccmd.tm.service_5_event import Service5TM
-from tmtccmd.tm.service_3_base import Service3Base
 from tmtccmd.tm.service_3_housekeeping import Service3TM
+
+from tmtccmd.tm.definitions import TelemetryListT, TelemetryQueueT
+from tmtccmd.tm.service_5_event import Service5TM
 from tmtccmd.tm.service_8_functional_cmd import Service8TM
-from tmtccmd.tm.service_23_file_mgmt import Service23TM
