+++ src/tmtccmd/tm/handler.py
@@ -1,7 +1,7 @@
 from tmtccmd.tm.definitions import TmTypes
 from spacepackets.ecss.tm import PusTelemetry
 from tmtccmd.tm.service_5_event import Service5TM
-from tmtccmd.tm.service_1_verification import Service1TM
+from tmtccmd.pus.service_1_verification import Service1TM
 from tmtccmd.pus.service_17_test import Service17TMExtended
 from tmtccmd.utility.logger import get_console_logger
 from tmtccmd.utility.tmtc_printer import TmTcPrinter
