+++ src/tmtccmd/tm/service_20_parameters.py
@@ -2,10 +2,11 @@ from __future__ import annotations
 import os
 import struct
 
+from spacepackets.ecss.tm import CdsShortTimestamp, PusVersion, PusTelemetry
+
 from tmtccmd.pus.service_list import PusServices
 from tmtccmd.pus.obj_id import ObjectId
-from spacepackets.ecss.tm import CdsShortTimestamp, PusVersion
-from tmtccmd.tm.base import PusTmInfoBase, PusTmBase, PusTelemetryExtended
+from tmtccmd.tm.base import PusTmInfoBase, PusTmBase
 from tmtccmd.utility.logger import get_console_logger
 
 LOGGER = get_console_logger()
