+++ src/tmtccmd/utility/tmtc_printer.py
@@ -5,7 +5,8 @@ import enum
 from typing import cast
 
 from spacepackets.ecss.tc import PusTelecommand
-from tmtccmd.tm import Service8TM, Service5TM
+from tmtccmd.tm.service_8_functional_cmd import Service8TM
+from tmtccmd.tm.service_5_event import Service5TM
 from tmtccmd.pus.service_list import PusServices
 from tmtccmd.tm.base import PusTmInfoInterface, PusTmInterface
 from tmtccmd.pus.service_8_func_cmd import Srv8Subservices
